Rename: contract 상호작용 함수명 변경
