--- client/src/contract/callPurchase.js
@@ -1,31 +1,0 @@
-import { contractGetterSDK } from "./contractCall";
-
-export const callPurchaseDeposit = async (signer, productId, buyerId) => {
-  try {
-    const contract = await contractGetterSDK(signer);
-    const data = await contract.call(
-      "purchaseAmountDeposit",
-      [productId, buyerId],
-      {
-        gasLimit: 1000000,
-      }
-    );
-    console.log(data);
-    return data;
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-export const callRelease = async (signer, productId) => {
-  try {
-    const contract = await contractGetterSDK(signer);
-    const data = await contract.call("purchaseConfirmation", [productId], {
-      gasLimit: 1000000,
-    });
-    console.log(data);
-    return data;
-  } catch (error) {
-    console.log(error);
-  }
-};
