+++ client/src/contract/contractReadCall.js
@@ -1,0 +1,1 @@
+import { contractGetterSDK } from "./contract";
